Added (partial) support for minecarts

+++ src/main/java/me/bteuk/converterplugin/Plugin.java
@@ -7,6 +7,7 @@ import org.bukkit.World;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.json.simple.JSONArray;
+import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
 
@@ -129,7 +130,7 @@ public class Plugin extends JavaPlugin {
 
                     if (!converter.isRunning() && !converterQueue.isEmpty()) {
 
-                        //Set converter to running, this prevent multiple conversion running at once.
+                        //Set converter to running, this prevents multiple conversion running at once.
                         converter.setRunning(true);
 
                         //Get the first region in the set.
@@ -140,28 +141,40 @@ public class Plugin extends JavaPlugin {
 
                             getLogger().info("Starting conversion of " + newFile);
 
-                            //Get the array of json objects.
-                            JSONArray jsonArray = (JSONArray) parser.parse(reader);
+                            Object rawJsonObject = parser.parse(reader);
+                            JSONObject jsonObject = new JSONObject();
 
-                            Bukkit.getScheduler().runTaskAsynchronously(this, () -> {
+                            if(rawJsonObject instanceof JSONObject) {
+                                //Get the json object that contains blocks (and entities)
+                                jsonObject = (JSONObject) rawJsonObject;
+                            }else {
+                                //Put array of json blocks into json object (for backwards compatibility)
+                                JSONArray jsonArray = (JSONArray) rawJsonObject;
+                                jsonObject.put("block", jsonArray);
+                            }
 
-                                CompletableFuture<Void> converterTask = converter.convert(jsonArray);
+                            if(!jsonObject.isEmpty()) {
+                                JSONObject finalJsonObject = jsonObject;
+                                Bukkit.getScheduler().runTaskAsynchronously(this, () -> {
 
-                                converterTask.thenRun(() -> {
-                                    getLogger().info("Converted file " + newFile);
+                                    CompletableFuture<Void> converterTask = converter.convert(finalJsonObject);
 
-                                    //If the converter has converted the region successfully, remove the file from the queue and set isRunning to false to unlock the converter.
-                                    //Additionally delete the file, so it can't be converted multiple times.
-                                    if (newFile.delete()) {
-                                        getLogger().info("Deleted " + newFile.getName());
-                                    } else {
-                                        getLogger().info("Failed to delete " + newFile.getName());
-                                    }
+                                    converterTask.thenRun(() -> {
+                                        getLogger().info("Converted file " + newFile);
 
-                                    converterQueue.remove(newFile);
-                                    converter.setRunning(false);
+                                        //If the converter has converted the region successfully, remove the file from the queue and set isRunning to false to unlock the converter.
+                                        //Additionally delete the file, so it can't be converted multiple times.
+                                        if (newFile.delete()) {
+                                            getLogger().info("Deleted " + newFile.getName());
+                                        } else {
+                                            getLogger().info("Failed to delete " + newFile.getName());
+                                        }
+
+                                        converterQueue.remove(newFile);
+                                        converter.setRunning(false);
+                                    });
                                 });
-                            });
+                            }
 
                         } catch (IOException | ParseException ex) {
                             ex.printStackTrace();
